+++ crosshatch/KnownAssets/SBC.ts
@@ -13,13 +13,6 @@ export const SBC = {
         version: "1",
         adapters: [Erc3009.Erc3009Adapter, Permit2.Permit2Adapter],
       },
-      72344: {
-        asset: EvmAsset.EvmAsset.make("0x33ad9e4BD16B69B5BFdED37D8B5D9fF9aba014Fb"),
-        decimals: 6,
-        name: "Stable Coin",
-        version: "1",
-        adapters: [Erc3009.Erc3009Adapter, Permit2.Permit2Adapter],
-      },
     },
   },
 } as const satisfies PhysicalAsset
