+++ crosshatch/KnownAssets/USDC.ts
@@ -13,13 +13,6 @@ export const USDC = {
         version: "2",
         adapters: [Erc3009.Erc3009Adapter, Permit2.Permit2Adapter],
       },
-      51: {
-        asset: EvmAsset.EvmAsset.make("0xb5AB69F7bBada22B28e79C8FFAECe55eF1c771D4"),
-        decimals: 6,
-        name: "USDC",
-        version: "2",
-        adapters: [Erc3009.Erc3009Adapter, Permit2.Permit2Adapter],
-      },
       137: {
         asset: EvmAsset.EvmAsset.make("0x3c499c542cEF5E3811e1192ce70d8cC03d5c3359"),
         decimals: 6,
@@ -45,20 +38,6 @@ export const USDC = {
         asset: EvmAsset.EvmAsset.make("0xaf88d065e77c8cC2239327C5EDb3A432268e5831"),
         decimals: 6,
         name: "USD Coin",
-        version: "2",
-        adapters: [Erc3009.Erc3009Adapter, Permit2.Permit2Adapter],
-      },
-      421614: {
-        asset: EvmAsset.EvmAsset.make("0x75faf114eafb1BDbe2F0316DF893fd58CE46AA4d"),
-        decimals: 6,
-        name: "USD Coin",
-        version: "2",
-        adapters: [Erc3009.Erc3009Adapter, Permit2.Permit2Adapter],
-      },
-      84532: {
-        asset: EvmAsset.EvmAsset.make("0x036CbD53842c5426634e7929541eC2318f3dCF7e"),
-        decimals: 6,
-        name: "USDC",
         version: "2",
         adapters: [Erc3009.Erc3009Adapter, Permit2.Permit2Adapter],
       },
