+++ crosshatch/index.ts
@@ -1,5 +1,4 @@
 export * as Amount from "./Amount.ts"
-export * as Payments from "./Payments.ts"
 export * as Payer from "./Payer.ts"
 export * from "./Stage.ts"
 export * as Requirements from "./Requirements.ts"
@@ -15,7 +14,6 @@ export * as ResourceInfo from "./ResourceInfo.ts"
 export * as Version from "./Version.ts"
 export * as Facilitator from "./Facilitator/Facilitator.ts"
 export * as Extension from "./Extension.ts"
-export * from "./extensions/PaymentId.ts"
 export * as Mnemonic from "./Mnemonic.ts"
 export * as Http402 from "./Http402/Http402.ts"
 export * as KnownAssets from "./KnownAssets/KnownAssets.ts"
