+++ crosshatch/package.json
@@ -24,7 +24,7 @@
     "./Browser": "./Browser/Browser.ts",
     "./Crypto": "./Crypto/Crypto.ts",
     "./Evm": "./Evm/Evm.ts",
-    "./Extensions": "./Extension.ts",
+    "./Extensions": "./Extensions/Extensions.ts",
     "./Facade": "./Browser/Facade/Facade.ts",
     "./Ramp": "./Ramp/Ramp.ts",
     "./Svm": "./Svm/Svm.ts",
@@ -37,7 +37,7 @@
       "./Browser": "./dist/Browser/Browser.js",
       "./Crypto": "./dist/Crypto/Crypto.js",
       "./Evm": "./dist/Evm/Evm.js",
-      "./Extensions": "./dist/Extension/Extension.ts",
+      "./Extensions": "./dist/Extensions/Extensions.js",
       "./Facade": "./dist/Browser/Facade/Facade.js",
       "./Ramp": "./dist/Ramp/Ramp.js",
       "./Svm": "./dist/Svm/Svm.js",
